this is dylan
I added the duration in game played for snake and 2048 game by changing the data model of gamesplayed's variable hours into time elapsed and changed the data type. On end of game activity, the duration is added automatically.

For the score system, since 2048 is a highscore based game, only the high score is saved and is saved in the globalVar.kt file under the variable name twentyScore.

For the snake game, after loss of game the score is saved into the globalVar.kt file under the name snakeScoreList which is an array of intergers.